fixing some tests
